proj6/pyproject.toml
[tool.poetry]
name = "proj6"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
numba = "^0.56.4"
numpy = "^1.23.5"
pillow = "^9.3.0"
tqdm = "^4.64.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
